Timer and gameOverMenu separation of concerns
